requirements.txt
Flask==2.2.5
Flask_Bcrypt==1.0.1
Flask_Login==0.6.3
Flask_Mail==0.9.1
flask_sqlalchemy==3.1.1
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.1
torch==2.1.2
torchvision==0.16.2
